Adding add.jsp for sessioncCourse

+++ src/main/webapp/WEB-INF/views/body/clients/all.jsp
@@ -25,7 +25,7 @@
 			      </td>
 			      <td>
 			       <spring:url value="/clients/deleteClient/${client.getId()}" var="deleteURL" />
-			       <a class="btn btn-danger" href="${deleteURL}" role="button" >Delete</a>
+			       <a class="btn btn-danger" onclick="return confirm('Are you sure, you want to delete this item?');" href="${deleteURL}" role="button" >Delete</a>
 			      </td>
 		     </tr>
 		    </c:forEach>
